Making the time zone work again
The time zone was being validated for presence and it wasn't getting passed in on user creation, so it failed on saving the model. Unfortunately, that failure was quiet, so the user just got redirected to the home page without any errors.

This commit fixes that issue by:
* Checking for login failures and flashing an error to the user, as well as logging the problem
* Removing validation for time one

It also adds:
* Image URLs for the user profiles
* Correct updating of the povider and UID attributes for the user