Adds cancel ticket button
Adds a button for the current user to cancel their ticket. Fixes #2 (badly). It
destroys the ticket, so no trace is left of it. Should really add some
functionality to let the user leave a comment and add a canceled status rather
than taking the heavy-handed approach.

Also adds better randomization to chat and makes the input field autofocused.